OOPFinalProject/MyFiles/requirements.txt
beautifulsoup4==4.13.4
pandas==2.3.0
Pillow==11.2.1
python-dotenv==1.1.0
Requests==2.32.4
selenium==4.33.0
